% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actions.per.dz.plot.R
\name{actions.per.dz.plot}
\alias{actions.per.dz.plot}
\title{Create a bar graph of FDA Approvals Per Disease}
\usage{
actions.per.dz.plot(data)
}
\arguments{
\item{data}{A data frame of FDA cutaneous oncology data}
}
\value{
a data frame
}
\description{
`actions.per.dz.plot()` Create a bar graph of FDA Approvals Per Disease
}
