% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.dt.to.plot.df.R
\name{convert.dt.to.plot.df}
\alias{convert.dt.to.plot.df}
\title{Converts the dt from the FDA cutaneous oncology into one that can be used by the `fda.plot()` function}
\usage{
convert.dt.to.plot.df(data)
}
\arguments{
\item{data}{A data frame of FDA cutaneous oncology data}
}
\value{
a data frame
}
\description{
`convert.dt.to.plot.df()` converts the dt from the FDA cutaneous oncology into one that can be used by the `fda.plot()` function
}
