% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda_actions_per_disease_plot.R
\name{fda_actions_per_disease_plot}
\alias{fda_actions_per_disease_plot}
\title{Create a bar graph of FDA Approvals Per Disease}
\usage{
fda_actions_per_disease_plot(.data = skincancerRx_data)
}
\arguments{
\item{.data}{A data frame of FDA cutaneous oncology data}
}
\value{
a data frame
}
\description{
`fda_actions_per_disease_plot()` Create a bar graph of FDA approvals per type of skin cancer
}
