% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda_approval_timeline_df.R
\name{fda_approval_timeline_df}
\alias{fda_approval_timeline_df}
\title{Creates a data frame that can be used by fda_appropval_timeline_plot to generate an interactive plot}
\usage{
fda_approval_timeline_df(data)
}
\arguments{
\item{data}{A data frame of FDA cutaneous oncology data}
}
\value{
a data frame
}
\description{
`fda_approval_timeline_df()` Creates a data frame that can be used by fda_appropval_timeline_plot to generate an interactive plot
}
