% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda_approval_timeliner_df.R
\name{fda_approval_timeliner_df}
\alias{fda_approval_timeliner_df}
\title{Create a df of the FDA Data that can be used for TimelineR}
\usage{
fda_approval_timeliner_df(.data = skincancerRx_data)
}
\arguments{
\item{.data}{fda data set, defaults to the skincancerRx_data dataset}
}
\value{
A ggplotly Timeline for a given data set
}
\description{
`fda_approval_timeliner_df()` takes the FDA data set and wrangles it into a DF that can then be used for TimelineR
}
