% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda_approval_timeliner_plot.R
\name{fda_approval_timeliner_plot}
\alias{fda_approval_timeliner_plot}
\title{Create a plotly interactive timeline of the FDA approvals for skin cancer}
\usage{
fda_approval_timeliner_plot(
  .data = fda_approval_timeliner_df(),
  .title = "<b>FDA Approvals in Cutaneous Oncology<b>",
  .startbuffer = 1000,
  .endbuffer = 1000,
  .labelwrap = 100,
  .textbuffer = 0.5,
  .topboundary = 100,
  .bottomboundary = 0,
  .geomtextsize = 4,
  .legendtextsize = 12,
  .hoverlabeltextsize = 20,
  .xaxistickfontsize = 14,
  .legend = TRUE
)
}
\arguments{
\item{.data}{is a data frame down stream of fda_data_for_timeliner. It defaults to the embedded dataset transformed by fda_data_for_timeliner, but allows for customization if the default is not used}

\item{.title}{Character string for the title of the character string}

\item{.startbuffer}{numeric value that represents the number of days prior to the first approved therapy on the data visualization}

\item{.endbuffer}{numeric value that represents the number of days after to the last approved therapy on the data visualization}

\item{.labelwrap}{numeric value that represents the length in charcters that will trigger a wrap of the text of each drug on the data visualization}

\item{.textbuffer}{numeric value that represents the distance away from the vetical line segment on the timeline}

\item{.topboundary}{numeric value that represents the distance from the top of data visualization to the title}

\item{.bottomboundary}{numeric value that represents the distance from the bottom of the data visualization to the legend}

\item{.geomtextsize}{numeric value of the geom_text font size}

\item{.legendtextsize}{numeric value of the legend text font size}

\item{.hoverlabeltextsize}{numeric value of the hover text font size}

\item{.xaxistickfontsize}{numeric value of the x axis text font size}

\item{.legend}{TRUE or FALSE to show legend or not}
}
\value{
A ggplotly Timeline of the FDA data set
}
\description{
fda_approval_timeliner_plot()` takes the aggregated data frame from fda_data_for_timeliner to produce a plotly data visualization the FDA approvals for skin cancer
}
\examples{
fda_skin_cancer_data |> 
 filter(Dz == "Melanoma") |> 
 fda_data_for_timeliner() |> 
 fda_approval_timeliner_plot()
}
