% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda_approval_timeseries_df.R
\name{fda_approval_timeseries_df}
\alias{fda_approval_timeseries_df}
\title{Creates a data frame that can be used by fda_approval_timeseries_plot to generate an interactive plot}
\usage{
fda_approval_timeseries_df()
}
\value{
a data frame
}
\description{
`fda_approval_timeseries_df()` Creates a data frame that can be used by fda_appropval_timeseries_plot to generate an interactive plot
}
