% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDA_Data_for_TimelineR.R
\name{fda_data_for_timeliner}
\alias{fda_data_for_timeliner}
\title{Create a df of the FDA Data that can be used for TimelineR}
\usage{
fda_data_for_timeliner()
}
\arguments{
\item{.data}{fda data set}
}
\value{
A ggplotly Timeline for a given data set
}
\description{
`fda.data.for.timeliner()` takes the FDA data set and wrangles it into a DF that can then be used for TimelineR
}
