% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda_skin_cancer_data.R
\name{fda_skin_cancer_data}
\alias{fda_skin_cancer_data}
\title{fda_skin_cancer_data}
\usage{
fda_skin_cancer_data()
}
\value{
data of FDA approvals for skin cancer
}
\description{
fda_skin_cancer_data
}
