% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_rate_df.R
\name{response_rate_df}
\alias{response_rate_df}
\title{Creates a data frame that can be used by response_rate_plot to generate an interactive plot}
\usage{
response_rate_df(data)
}
\arguments{
\item{data}{A data frame of FDA cutaneous oncology data}
}
\value{
a data frame
}
\description{
`response_rate_df()` Creates a data frame that can be used by response_rate_plot to generate an interactive plot trials
}
