% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.rx.skin.cancer.R
\name{table.rx.skin.cancer}
\alias{table.rx.skin.cancer}
\title{Create a data frame of drugs with an indication for skin cancer}
\usage{
table.rx.skin.cancer()
}
\value{
an interactive plot
}
\description{
`table.rx.skin.cancer()` Creates a table of fda approvals in skin cancer
}
