% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_rx_skin_cancer.R
\name{table_rx_skin_cancer}
\alias{table_rx_skin_cancer}
\title{Create a data frame of drugs with an indication for skin cancer}
\usage{
table_rx_skin_cancer(.data = skincancerRx_data)
}
\arguments{
\item{.data}{A data frame of unprocessed FDA appovals in skin cancer. Defaults to the embedded data set}
}
\value{
a data frame
}
\description{
`table_rx_skin_cancer()` Creates a table of fda approvals in skin cancer
}
