% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimelineR_plot.R
\name{timeliner_plot}
\alias{timeliner_plot}
\title{Create a plotly interactive timeline of a patient journey from a tumor registry}
\usage{
timeliner_plot(
  .data,
  Title = "<b>FDA Approvals in Cutaneous Oncology<b>",
  StartBuffer = 1000,
  EndBuffer = 1000,
  LabelWrap = 100,
  TextBuffer = 0.5,
  TopBoundary = 100,
  BottomBoundary = 0,
  Legend = TRUE
)
}
\arguments{
\item{data}{is a data frame down stream of FDA_Data_for_TimelineR}
}
\value{
A ggplotly Timeline for a given data set
}
\description{
`timeliner_plot()` takes the aggregated data frame from FDA_Data_for_TimelineR to produce a plotly data visualization of a patient journey
}
