% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda_approval_timeseries_plot.R
\name{fda_approval_timeseries_plot}
\alias{fda_approval_timeseries_plot}
\title{Creates an interactive plot of fda approvals in skin cancer overtime}
\usage{
fda_approval_timeseries_plot(
  .data,
  Title = "<b>FDA Approvals in Cutaneous Oncology<b>",
  StartDate = "1948-01-01",
  EndDate = Sys.Date(),
  Legend = TRUE
)
}
\arguments{
\item{.data}{DF downstream of fda_approval_timeseries_df.}
}
\value{
an interactive plot
}
\description{
`fda_approval_timeseries_plot()` Creates an interactive plot of fda approvals in skin cancer overtime
}
