% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_rate_plot.R
\name{response_rate_plot}
\alias{response_rate_plot}
\title{Creates an interactive plot of response rates for single arm trials}
\usage{
response_rate_plot(data)
}
\arguments{
\item{data}{A data frame of response rates of therapies Required.}
}
\value{
an interactive plot
}
\description{
`response_rate_plot()` Creates an interactive plot of response rates for single arm trials
}
